fix: sample app ui bugs (#2854)
* fix: sample app ui bugs

* fix: create new channel bug

* fix: navigation go back bug
