+++ examples/SampleApp/src/components/ScreenHeader.tsx
@@ -64,7 +64,13 @@ export const BackButton: React.FC<{
   return (
     <TouchableOpacity
       onPress={() => {
-        navigation.goBack();
+        if (!navigation.canGoBack()) {
+          // if no previous screen was present in history, go to the list screen
+          // this can happen when opened through push notification
+          navigation.reset({ index: 0, routes: [{ name: 'HomeScreen' }] });
+        } else {
+          navigation.goBack();
+        }
         if (onBack) {
           onBack();
         }
@@ -165,6 +171,7 @@ export const ScreenHeader: React.FC<ScreenHeaderProps> = (props) => {
                       color: black,
                     },
                   ]}
+                  numberOfLines={1}
                 >
                   {titleText}
                 </Text>
@@ -182,6 +189,7 @@ export const ScreenHeader: React.FC<ScreenHeaderProps> = (props) => {
                     color: grey,
                   },
                 ]}
+                numberOfLines={1}
               >
                 {subtitleText}
               </Text>
